pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "ur_l"
version = "0.1.0"
description = "ur_l - URL Shortener by adamwojt"
authors = ["Adam Wojtczak <adam1edinburgh@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"

coverage = { version = "5.0.3", extras = ["toml"]}
django = "3.1"
django-redis = "4.12.1"
djangorestframework = "3.11.1"
fakeredis = "1.4.2"
gunicorn = "^20.0.4"
psycopg2-binary = "2.8.5"
pytest = "5.3.5"
pytest-cov = "2.10.0"
pytest-django = "3.9.0"
redis = "3.5.3"
coreapi = "2.3.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
isort = { version = "^4.3.21", extras = ["pyproject"]}
safety = "1.9.0"

[tool.black]
# https://github.com/psf/black
line-length = 88
target_version = ['py38']
exclude = '''
(
  /(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | htmlcov
    | venv
    | .venv
    | migrations
  )/
)
'''

[tool.isort]
# https://github.com/timothycrosley/isort
# https://github.com/timothycrosley/isort/wiki/isort-Settings
line_length = 88
indent = '    '
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0

[tool.coverage]
# https://github.com/nedbat/coveragepy
    [tool.coverage.run]

    source = ["app"]
    branch = true
    omit = ['']

    [tool.coverage.report]
    exclude_lines = [
        "# noqa",
        "raise NotImplementedError"
    ]
